refactor(app): move the git view and code review state into ainb-app
GitViewState and its file tree, markdown and commit types move to
ainb_app::components::git_view. The code review model and parser move
whole, and the review UI state (CodeReviewUi) moves with the keyboard and
mouse helpers that act on it. The draw functions and the syntax colour
bridge stay in ainb-core. ainb-app picks up similar for the parser.

Two ratatui types leave the state:
- GitFileStatus::color becomes git_view::status_color in ainb-core.
- CodeReviewUi's last drawn sidebar rectangle, which a click is
  hit-tested against, becomes a Cell<Area>; render converts the Rect it
  drew into. Both renderer-written cells become pub so ainb-core's render
  can still write them.
